+++ packages/gatsby-plugin-sharp/safe-sharp.d.ts
@@ -1,0 +1,2 @@
+import sharp from "sharp"
+export default sharp
